Create Business::CPI::Account classes
Originally, Business::CPI used to manage accounts in the gateway only in
a purchase context. That means that, when building the cart, and making
a payment, we sent the buyer information, built in the
Business::CPI::Buyer class. But gateways also provide means in their
API's to create accounts, not only for clients, but also resellers,
representants, and so forth.

So what this commit provides is a way to represent those accounts when
the driver intends to manage accounts in the gateway.

The next step would be to make Business::CPI::Buyer extend
Business::CPI::Account.